% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars_price.R
\name{cars_price}
\alias{cars_price}
\title{Calculats car price}
\usage{
cars_price(cars, price)
}
\arguments{
\item{cars}{Number of cars}

\item{price}{Price per car}
}
\value{
returns total cost
}
\description{
Calculats car price
}
\examples{
cars_price(cars = 2, price = 20000)
}
